+++ assets/asset-discord/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-discord
 
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- update
+
 ## 1.0.0-beta.1
 
 ### Patch Changes