+++ plugins/plugin-discord/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/plugin-discord
 
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.3
+  - @pipelab/shared@1.0.0-beta.2
+
 ## 1.0.0-beta.2
 
 ### Patch Changes