+++ plugins/plugin-steam/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/plugin-steam
 
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.3
+
 ## 1.0.0-beta.2
 
 ### Patch Changes